# backend/tests/contracts/openapi-baseline.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlanAdminPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlanOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages"
        ],
        "title": "PlanAdminPage",
        "type": "object"
      },
      "PlanIn": {
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "code": {
            "maxLength": 20,
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9_-]*$",
            "title": "Code",
            "type": "string"
          },
          "label_en": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Label En",
            "type": "string"
          },
          "label_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Label Id",
            "type": "string"
          },
          "months": {
            "maximum": 36.0,
            "minimum": 1.0,
            "title": "Months",
            "type": "integer"
          },
          "order": {
            "default": 1,
            "maximum": 999.0,
            "minimum": 1.0,
            "title": "Order",
            "type": "integer"
          },
          "price": {
            "minimum": 0.0,
            "title": "Price",
            "type": "integer"
          }
        },
        "required": [
          "code",
          "label_id",
          "label_en",
          "months",
          "price"
        ],
        "title": "PlanIn",
        "type": "object"
      },
      "PlanOut": {
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "code": {
            "maxLength": 20,
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9_-]*$",
            "title": "Code",
            "type": "string"
          },
          "created_at": {
            "default": "",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label_en": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Label En",
            "type": "string"
          },
          "label_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Label Id",
            "type": "string"
          },
          "months": {
            "maximum": 36.0,
            "minimum": 1.0,
            "title": "Months",
            "type": "integer"
          },
          "order": {
            "default": 1,
            "maximum": 999.0,
            "minimum": 1.0,
            "title": "Order",
            "type": "integer"
          },
          "price": {
            "minimum": 0.0,
            "title": "Price",
            "type": "integer"
          },
          "updated_at": {
            "default": "",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "code",
          "label_id",
          "label_en",
          "months",
          "price",
          "id"
        ],
        "title": "PlanOut",
        "type": "object"
      },
      "PlannerAnalyticsEventIn": {
        "additionalProperties": false,
        "description": "Minimal, consented planner-funnel event. It deliberately has no story field.",
        "properties": {
          "anonymous_session_id": {
            "maxLength": 80,
            "minLength": 16,
            "title": "Anonymous Session Id",
            "type": "string"
          },
          "event_id": {
            "maxLength": 80,
            "minLength": 16,
            "title": "Event Id",
            "type": "string"
          },
          "event_type": {
            "enum": [
              "planner_story_submitted",
              "planner_step_shown",
              "planner_step_completed",
              "planner_generated"
            ],
            "title": "Event Type",
            "type": "string"
          },
          "step": {
            "enum": [
              "story",
              "basics",
              "interests",
              "result"
            ],
            "title": "Step",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "event_type",
          "step",
          "anonymous_session_id"
        ],
        "title": "PlannerAnalyticsEventIn",
        "type": "object"
      },
...
      "TripPlanIn": {
        "properties": {
          "budget": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget"
          },
          "budget_style": {
            "anyOf": [
              {
                "enum": [
                  "budget",
                  "mid_range",
                  "luxury"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Style"
          },
          "days": {
            "maximum": 14.0,
            "minimum": 1.0,
            "title": "Days",
            "type": "integer"
          },
          "extra_context": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "",
            "title": "Extra Context"
          },
          "interests": {
            "items": {
              "enum": [
                "adventure",
                "beach",
                "camping",
                "culinary",
                "culture",
                "hotel",
                "hotspring",
                "island",
                "lake",
                "mountain",
                "nature",
                "tea",
                "viewpoint",
                "waterfall"
              ],
              "type": "string"
            },
            "maxItems": 14,
            "title": "Interests",
            "type": "array"
          },
          "lang": {
            "default": "id",
            "enum": [
              "id",
              "en"
            ],
            "title": "Lang",
            "type": "string"
          },
          "preferred_destination_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Preferred Destination Ids",
            "type": "array"
          },
          "previous_content": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "",
            "title": "Previous Content"
          }
        },
        "required": [
          "days"
        ],
        "title": "TripPlanIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Explore Wisata Sumut API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/premium/plans": {
      "get": {
        "operationId": "list_admin_plans_api_admin_premium_plans_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "all",
              "enum": [
                "all",
                "active",
                "inactive"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "order",
              "title": "Sort",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanAdminPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Admin Plans"
      },
      "post": {
        "operationId": "create_plan_api_admin_premium_plans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Plan"
      }
    },
    "/api/admin/premium/plans/{plan_id}": {
      "delete": {
        "operationId": "delete_plan_api_admin_premium_plans__plan_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Plan"
      },
      "put": {
        "operationId": "update_plan_api_admin_premium_plans__plan_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Plan"
      }
    },
...
    "/api/analytics/planner-events": {
      "post": {
        "description": "Store aggregate planner-funnel telemetry, never the user's planner story.",
        "operationId": "track_planner_event_api_analytics_planner_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlannerAnalyticsEventIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Track Planner Event"
      }
    },
...
    "/api/share/{slug}": {
      "get": {
        "operationId": "share_preview_page_api_share__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Share Preview Page"
      }
    },
    "/api/share/{slug}/image.png": {
      "get": {
        "operationId": "share_card_image_api_share__slug__image_png_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Share Card Image"
      }
    },
    "/api/trip-planner/stream": {
      "post": {
        "operationId": "trip_planner_stream_api_trip_planner_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripPlanIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trip Planner Stream"
      }
    },
...
  }
}